backend/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy>=1.21.0,<2
Pillow>=9.0.0
scikit-learn>=1.0.0
ftfy
regex
tqdm
clip @ git+https://github.com/openai/CLIP.git
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
websockets>=12.0
